config/requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
certifi==2021.10.8
charset-normalizer==2.0.7
conda==4.3.16
distlib==0.3.3
Django==4.0.1
django-ckeditor==6.2.0
django-debug-toolbar==3.2.4
django-js-asset==1.2.2
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
filelock==3.3.2
idna==3.3
isort==5.10.1
Pillow==9.0.0
pipenv==2021.5.29
platformdirs==2.4.0
pycosat==0.6.3
pytils==0.3
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
virtualenv==20.10.0
virtualenv-clone==0.5.7
